requirements.txt
discord.py
flask
gunicorn
pip install python-dateutil
python-dotenv
pytz
waitress